+++ changelog.md
@@ -1,4 +1,8 @@
 
+## 0.0.4
+* Add missing lang to porting lib attributes (temporary)
+* Add a config to show/hide attributes in the attributes screen
+
 ## 0.0.3
 * Reduce priority of mixin to fix tooltip rendering